Fetch serie and season images when adding to favorites
works pretty well. I am not sure if we should fetch images before we add
to favorites or add and fetch in background, i figured it would be extra
work and frankly, this works very well, even with a rate limit of 40,
it's HIGHLY unlikely anyone will be able to reach the rate limit while
adding shows when you factor in the delay of actually fetching all the
data from trakt and then getting images, I clicked as fast as I could
adding over 40 shows and nothing happened. Also thanks to a SECRET
trick, we can get all the season posters and serie images/fanart in a
single call .