Fetch images for trending
Update trending cache with preexisting image IDs, the URLs are generated
when they are added as it could have changed.

Added a possible method for updates to the trending list, when the
trendinglist updates and new series are added, if the serie was already
in the list, it will use it's existing poster and not fetch a new one,
if it doesn't exist it will attempt to fetch it.

Oh and, I set it down to 400 series cause I thought 500 was overkill
anyway but, sending 500 requests for images is probably not ideal.
Hopefully there shouldn't be too many image updates and trending only
updates every 7 days anyway.